http://jira.jboss.org/jira/browse/EXIN-235 Mozilla was replaced with xulrunner
git-svn-id: http://svn.jboss.org/repos/jbosstools/branches@2978 a97e2381-89e5-4abb-bab3-167db6db766c

+++ jbosstools_xulrunner/vpe/plugins/org.jboss.tools.vpe/src/org/jboss/tools/vpe/editor/util/TextUtil.java
@@ -252,8 +252,8 @@ public class TextUtil {
 		return sourceIndex;
 	}
 
-	public static int sourceInnerPosition(String visualText, int visualPosition) {
-		visualText = visualText.substring(0, Math.min(visualPosition, visualText.length()));
+	public static int sourceInnerPosition(String visualText, long visualPosition) {
+		visualText = visualText.substring(0, (int)Math.min(visualPosition, visualText.length()));
 		String sourceText = visualText.replaceAll(VISUAL_BREAK, SOURCE_BREAK);
 		return sourceText.length();
 	}
